requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
Django==3.1.7
django-environ==0.4.5
djangorestframework==3.12.2
gunicorn==20.0.4
html5lib==1.1
idna==2.10
json-table-schema==0.2.1
lxml==4.6.2
Pillow==8.1.2
python-dateutil==2.8.1
python-magic==0.4.22
pytz==2021.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.3
webencodings==0.5.1
xlrd==2.0.1